api-v2/recomendations/requirements.txt
Flask==2.3.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.2.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0